+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   fixed infinite message bug
+
 ## [0.0.9] - 2020-03-13
 
 ### Fixed